changed openssl to use sha256 instead of deprecated sha1

+++ beehive-cert/SSL/create_certificate_authority.sh
@@ -1,23 +1,22 @@
 #!/bin/bash
 
+SSL_DIR="/usr/lib/waggle/SSL"
+CA_DIR="${SSL_DIR}/waggleca"
 
-export SSL_DIR="/usr/lib/waggle/SSL"
+mkdir -p $CA_DIR
+#chmod 700 $SSL_DIR
+chmod 700 $CA_DIR
 
-# Begin constructing the Certificate Authority
-
-mkdir -p /usr/lib/waggle
-chmod 777 -R /usr/lib/waggle
-
-
-rm -rf ${SSL_DIR}/waggleca
-mkdir -p ${SSL_DIR}/waggleca
-
-cd ${SSL_DIR}/waggleca
+cd $CA_DIR
 
 # Make appropriate folders
-mkdir -p certs private
+mkdir certs private
 chmod 700 private
-echo 01 > serial
+
+if [ ! -f serial ]; then
+	echo 01 > serial
+fi
+
 touch index.txt
 
 # this is needed for "node" certificates. We may change that later.
@@ -25,9 +24,23 @@ echo "unique_subject = no" > index.txt.attr
 
 # Generate the root certificate
 
-openssl req -x509 -config /usr/lib/waggle/beehive-server/SSL/waggleca/openssl.cnf -newkey rsa:2048 -days 365 \
-	-out cacert.pem -outform PEM -subj /CN=waggleca/ -nodes
+if [ -f "$CA_DIR/private/cakey.pem" ]; then
+	echo "Error: CA key already exists."
+fi
+
+if [ -f "$CA_DIR/cacert.pem" ]; then
+	echo "Error: CA cert already exists."
+fi
+
+openssl req \
+	-x509 \
+	-config /usr/lib/waggle/beehive-server/beehive-cert/SSL/waggleca \
+	-newkey rsa:2048 \
+	-keyform PEM \
+	-nodes \
+	-days 365 \
+	-out cacert.pem \
+	-outform PEM \
+	-subj /CN=waggleca/
 
 openssl x509 -in cacert.pem -out cacert.cer -outform DER
-
-
